+++ accounts-line/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'storyteller:accounts-line',
   summary: 'Login service for LINE accounts',
-  version: '1.1.0',
+  version: '1.2.1',
   git: 'https://github.com/StorytellerCZ/meteor-accounts-line'
 });
 
@@ -12,11 +12,11 @@ Package.onUse(api => {
   // Export Accounts (etc) to packages using this one.
   api.imply('accounts-base', ['client', 'server']);
   api.use('accounts-oauth', ['client', 'server']);
-  api.use('storyteller:line-oauth@1.1.0');
-  api.imply('storyteller:line-oauth@1.1.0');
+  api.use('storyteller:line-oauth@1.2.0');
+  api.imply('storyteller:line-oauth@1.2.0');
 
   // If users use accounts-ui but not facebook-config-ui, give them a tip.
-  api.use(['accounts-ui', 'storyteller:line-config-ui@1.1.0'], ['client', 'server'], { weak: true });
+  api.use(['accounts-ui', 'storyteller:line-config-ui@1.1.1'], ['client', 'server'], { weak: true });
   api.addFiles('notice.js');
 
   api.addFiles('line.js');