+++ line-oauth/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 1.1.0
+## 1.2.0 - 2021-02-16
+* Added `botPrompt` and `uiLocales` options to client
+* Updated the README to document options and config parameters
+* Fixed multiple issues from 1.1.0
+
+## 1.1.0 - 2021-02-14
 * Bump minimum Meteor version to 1.9
 * Replace HTTP package with Fetch package